usize -> u32

+++ compiler/noirc_evaluator/src/ssa/ir/basic_block.rs
@@ -39,8 +39,8 @@ impl BasicBlock {
     }
 
     /// Retrieve the type of the given parameter
-    pub(crate) fn type_of_parameter(&self, parameter_index: usize) -> &Type {
-        &self.parameter_types[parameter_index]
+    pub(crate) fn type_of_parameter(&self, parameter_index: u32) -> &Type {
+        &self.parameter_types[parameter_index as usize]
     }
 
     /// Adds a parameter to this BasicBlock.
